Refactor dataedge for common parts with htsget
