+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/dataedge/rest/MetadataController.java
@@ -17,7 +17,7 @@ package eu.elixir.ega.ebi.dataedge.rest;
 
 
 import eu.elixir.ega.ebi.dataedge.config.VerifyMessage;
-import eu.elixir.ega.ebi.dataedge.config.VerifyMessageNew;
+import eu.elixir.ega.ebi.shared.config.VerifyMessageNew;
 import eu.elixir.ega.ebi.shared.dto.File;
 import eu.elixir.ega.ebi.dataedge.service.FileMetaService;
 import org.springframework.beans.factory.annotation.Autowired;
