+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/dataedge/service/internal/RemoteFileServiceImpl.java
@@ -19,8 +19,12 @@ import com.google.common.io.ByteStreams;
 import com.google.common.io.CountingOutputStream;
 import eu.elixir.ega.ebi.dataedge.config.*;
 import eu.elixir.ega.ebi.dataedge.dto.*;
-import eu.elixir.ega.ebi.dataedge.service.DownloaderLogService;
-import eu.elixir.ega.ebi.dataedge.service.FileInfoService;
+import eu.elixir.ega.ebi.shared.config.GeneralStreamingException;
+import eu.elixir.ega.ebi.shared.config.PermissionDeniedException;
+import eu.elixir.ega.ebi.shared.dto.DownloadEntry;
+import eu.elixir.ega.ebi.shared.dto.EventEntry;
+import eu.elixir.ega.ebi.shared.service.DownloaderLogService;
+import eu.elixir.ega.ebi.shared.service.FileInfoService;
 import eu.elixir.ega.ebi.dataedge.service.FileService;
 import eu.elixir.ega.ebi.shared.dto.File;
 import eu.elixir.ega.ebi.shared.dto.FileIndexFile;
