+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/shared/rest/StatsController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.elixir.ega.ebi.dataedge.rest;
+package eu.elixir.ega.ebi.shared.rest;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.RequestHeader;
