+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/shared/service/AuthenticationService.java
@@ -1,4 +1,4 @@
-package eu.elixir.ega.ebi.dataedge.service;
+package eu.elixir.ega.ebi.shared.service;
 
 
 import java.util.Collection;
