+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/shared/service/internal/AbstractDownloaderLogService.java
@@ -1,9 +1,9 @@
-package eu.elixir.ega.ebi.dataedge.service.internal;
+package eu.elixir.ega.ebi.shared.service.internal;
 
-import eu.elixir.ega.ebi.dataedge.dto.DownloadEntry;
-import eu.elixir.ega.ebi.dataedge.dto.EventEntry;
-import eu.elixir.ega.ebi.dataedge.service.AuthenticationService;
-import eu.elixir.ega.ebi.dataedge.service.DownloaderLogService;
+import eu.elixir.ega.ebi.shared.dto.DownloadEntry;
+import eu.elixir.ega.ebi.shared.dto.EventEntry;
+import eu.elixir.ega.ebi.shared.service.AuthenticationService;
+import eu.elixir.ega.ebi.shared.service.DownloaderLogService;
 import java.util.Calendar;
 import javax.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
