+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/shared/service/internal/PermissionsServiceImpl.java
@@ -1,15 +1,15 @@
-package eu.elixir.ega.ebi.dataedge.service.internal;
+package eu.elixir.ega.ebi.shared.service.internal;
 
 import static eu.elixir.ega.ebi.shared.Constants.FILEDATABASE_SERVICE;
 
-import eu.elixir.ega.ebi.dataedge.config.GeneralStreamingException;
-import eu.elixir.ega.ebi.dataedge.config.NotFoundException;
-import eu.elixir.ega.ebi.dataedge.config.PermissionDeniedException;
-import eu.elixir.ega.ebi.dataedge.config.VerifyMessageNew;
-import eu.elixir.ega.ebi.dataedge.dto.EventEntry;
-import eu.elixir.ega.ebi.dataedge.service.AuthenticationService;
-import eu.elixir.ega.ebi.dataedge.service.DownloaderLogService;
-import eu.elixir.ega.ebi.dataedge.service.PermissionsService;
+import eu.elixir.ega.ebi.shared.config.GeneralStreamingException;
+import eu.elixir.ega.ebi.shared.config.NotFoundException;
+import eu.elixir.ega.ebi.shared.config.PermissionDeniedException;
+import eu.elixir.ega.ebi.shared.config.VerifyMessageNew;
+import eu.elixir.ega.ebi.shared.dto.EventEntry;
+import eu.elixir.ega.ebi.shared.service.AuthenticationService;
+import eu.elixir.ega.ebi.shared.service.DownloaderLogService;
+import eu.elixir.ega.ebi.shared.service.PermissionsService;
 import eu.elixir.ega.ebi.shared.dto.FileDataset;
 import java.util.Collection;
 import java.util.HashSet;
@@ -17,7 +17,6 @@ import java.util.Iterator;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.GrantedAuthority;
